+++ CHANGES.rst
@@ -4,13 +4,15 @@ Changelog
 5.0 (unreleased)
 ----------------
 
-- Drop support for Zope 2, require Zope 4.0b5 now.
+- Drop support for Zope 2, require Zope >= 4.0b6 now.
+
+- Add support for Python 3.5 and 3.6.
 
 - Drop `ZServer` dependency.
 
-- Add IPubFailure event handler so it writes error log entries again.
+- Add ``IPubFailure`` event handler so it writes error log entries again.
 
-- Bring back Application initialization (creation of SiteErrorLog in the
+- Bring back Application initialization (creation of `SiteErrorLog` in the
   ZODB on first startup).